--- tinytodo-types/Cargo.toml
@@ -1,22 +1,0 @@
-[package]
-name = "tiny-todo-server"
-edition = "2021"
-version = "0.1.0"
-publish = false
-
-[dependencies]
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-tokio = { version = "1", features = ["full"] }
-warp = "=0.3.5"
-uuid = { version = "1.3.0", features = ["v4", "fast-rng", "macro-diagnostics"] }
-thiserror =  "1"
-itertools = "0.10.5"
-tracing = "0.1.37"
-tracing-subscriber = "0.3.17"
-lazy_static = "1.4.0"
-notify = { version = "5.1.0", default-features = false, features = ["macos_kqueue"] }
-
-[dependencies.cedar-policy]
-version = "3.0.0"
-path = "../../cedar/cedar-policy"
