--- tinytodo-types/entities.json
@@ -1,58 +1,0 @@
-{
-  "users": {
-    "User::\"kesha\"": {
-      "euid": "User::\"kesha\"",
-      "parents": [
-        "Application::\"TinyTodo\"",
-        "Team::\"temp\""
-      ]
-    },
-    "User::\"aaron\"": {
-      "euid": "User::\"aaron\"",
-      "parents": [
-        "Team::\"interns\"",
-        "Application::\"TinyTodo\""
-      ]
-    },
-    "User::\"emina\"": {
-      "euid": "User::\"emina\"",
-      "parents": [
-        "Application::\"TinyTodo\"",
-        "Team::\"admin\""
-      ]
-    },
-    "User::\"andrew\"": {
-      "euid": "User::\"andrew\"",
-      "parents": [
-        "Application::\"TinyTodo\"",
-        "Team::\"admin\"",
-        "Team::\"temp\""
-      ]
-    }
-  },
-  "teams": {
-    "Team::\"temp\"": {
-      "uid": "Team::\"temp\"",
-      "parents": [
-        "Application::\"TinyTodo\""
-      ]
-    },
-    "Team::\"admin\"": {
-      "uid": "Team::\"admin\"",
-      "parents": [
-        "Application::\"TinyTodo\""
-      ]
-    },
-    "Team::\"interns\"": {
-      "uid": "Team::\"interns\"",
-      "parents": [
-        "Application::\"TinyTodo\"",
-        "Team::\"temp\""
-      ]
-    }
-  },
-  "lists": {},
-  "app": {
-    "euid": "Application::\"TinyTodo\""
-  }
-}
