--- tinytodo-types/tinytodo.cedarschema.json
@@ -1,161 +1,0 @@
-{
-	"": {
-		"entityTypes": {
-			"Application": {},
-			"User": {
-				"memberOfTypes": [
-					"Team",
-					"Application"
-				],
-				"shape": {
-					"type": "Record",
-					"attributes": {
-						"name": {
-							"type": "String"
-						}
-					}
-				}
-			},
-			"Team": {
-				"memberOfTypes": [
-					"Team",
-					"Application"
-				]
-			},
-			"List": {
-				"memberOfTypes": [
-					"Application"
-				],
-				"shape": {
-					"type": "Record",
-					"attributes": {
-						"owner": {
-							"type": "Entity",
-							"name": "User"
-						},
-						"name": {
-							"type": "String"
-						},
-						"readers": {
-							"type": "Entity",
-							"name": "Team"
-						},
-						"editors": {
-							"type": "Entity",
-							"name": "Team"
-						},
-						"tasks": {
-							"type": "Set",
-							"element": {
-								"type": "Record",
-								"attributes": {
-									"name": {
-										"type": "String"
-									},
-									"id": {
-										"type": "Long"
-									},
-									"state": {
-										"type": "String"
-									}
-								}
-							}
-						}
-					}
-				}
-			}
-		},
-		"actions": {
-			"CreateList": {
-				"appliesTo": {
-					"principalTypes": [
-						"User"
-					],
-					"resourceTypes": [
-						"Application"
-					]
-				}
-			},
-			"GetList": {
-				"appliesTo": {
-					"principalTypes": [
-						"User"
-					],
-					"resourceTypes": [
-						"List"
-					]
-				}
-			},
-			"UpdateList": {
-				"appliesTo": {
-					"principalTypes": [
-						"User"
-					],
-					"resourceTypes": [
-						"List"
-					]
-				}
-			},
-			"DeleteList": {
-				"appliesTo": {
-					"principalTypes": [
-						"User"
-					],
-					"resourceTypes": [
-						"List"
-					]
-				}
-			},
-			"GetLists": {
-				"appliesTo": {
-					"principalTypes": [
-						"User"
-					],
-					"resourceTypes": [
-						"Application"
-					]
-				}
-			},
-			"CreateTask": {
-				"appliesTo": {
-					"principalTypes": [
-						"User"
-					],
-					"resourceTypes": [
-						"List"
-					]
-				}
-			},
-			"UpdateTask": {
-				"appliesTo": {
-					"principalTypes": [
-						"User"
-					],
-					"resourceTypes": [
-						"List"
-					]
-				}
-			},
-			"DeleteTask": {
-				"appliesTo": {
-					"principalTypes": [
-						"User"
-					],
-					"resourceTypes": [
-						"List"
-					]
-				}
-			},
-			"EditShare": {
-				"appliesTo": {
-					"principalTypes": [
-						"User"
-					],
-					"resourceTypes": [
-						"List"
-					]
-				}
-			}
-		}
-	}
-}
